+++ bundle/src/main/java/com/adobe/acs/commons/marketo/client/impl/MarketoClientImpl.java
@@ -24,23 +24,13 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Dictionary;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-
-import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.http.auth.AUTH;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.osgi.service.component.annotations.Component;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.adobe.acs.commons.marketo.MarketoClientConfiguration;
 import com.adobe.acs.commons.marketo.client.MarketoClient;
@@ -49,6 +39,29 @@ import com.adobe.acs.commons.marketo.client.MarketoField;
 import com.adobe.acs.commons.marketo.client.MarketoForm;
 import com.adobe.acs.commons.marketo.client.MarketoResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.http.HttpHost;
+import org.apache.http.auth.AUTH;
+import org.apache.http.auth.AuthScope;
+import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.conn.routing.HttpRoutePlanner;
+import org.apache.http.impl.client.BasicCredentialsProvider;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.impl.conn.DefaultProxyRoutePlanner;
+import org.osgi.framework.InvalidSyntaxException;
+import org.osgi.service.cm.Configuration;
+import org.osgi.service.cm.ConfigurationAdmin;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Implementation of the MarketoClient using the REST API.
@@ -62,8 +75,72 @@ public class MarketoClientImpl implements MarketoClient {
 
   private ObjectMapper mapper = new ObjectMapper();
 
+  private HttpClientBuilder clientBuilder = HttpClients.custom();
+
+  @Reference
+  protected ConfigurationAdmin configAdmin;
+
+  @Activate
+  public void activate() {
+    Configuration[] configs;
+    try {
+      configs = configAdmin.listConfigurations("(service.factoryPid=org.apache.http.proxyconfigurator)");
+
+      for (Configuration config : configs) {
+        log.info("Evaluating proxy configuration: {}", config.getPid());
+
+        Dictionary<String, Object> properties = config.getProperties();
+        if (isEnabled(properties)) {
+          String host = String.class.cast(properties.get("proxy.host"));
+          int port = Integer.class.cast(properties.get("proxy.port"));
+
+          log.debug("Using proxy host: {}", host);
+          HttpHost proxyhost = new HttpHost(host, port);
+          HttpRoutePlanner routePlanner = new DefaultProxyRoutePlanner(proxyhost);
+          clientBuilder.setRoutePlanner(routePlanner);
+
+          String user = String.class.cast(properties.get("proxy.user"));
+          String password = String.class.cast(properties.get("proxy.password"));
+          if (StringUtils.isNotBlank(user)) {
+            log.debug("Using proxy authentication with user: {}", user);
+            BasicCredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+            credentialsProvider.setCredentials(new AuthScope(host, port),
+                new UsernamePasswordCredentials(user, password));
+            clientBuilder.setDefaultCredentialsProvider(credentialsProvider);
+          }
+
+        } else {
+          log.debug("Proxy configuration not enabled");
+        }
+
+      }
+    } catch (IOException | InvalidSyntaxException e) {
+      log.error("Failed to load proxy configuration", e);
+    }
+
+  }
+
+  private boolean isEnabled(Dictionary<String, Object> properties) {
+    if (properties == null) {
+      return false;
+    }
+
+    Object value = properties.get("proxy.enabled");
+    if (!(value instanceof Boolean) || !(Boolean.class.cast(value))) {
+      return false;
+    }
+
+    String host = String.class.cast(properties.get("proxy.host"));
+    if (StringUtils.isBlank(host)) {
+      return false;
+    }
+    return true;
+  }
+
   protected @Nonnull String getApiResponse(@Nonnull String url, String bearerToken) throws IOException {
-    try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
+
+    try (CloseableHttpClient httpClient = clientBuilder.build()) {
+      log.debug("Sending request to: {}", url);
       HttpGet httpGet = new HttpGet(url);
       if (StringUtils.isNotBlank(bearerToken)) {
         httpGet.setHeader(AUTH.WWW_AUTH_RESP, "Bearer " + bearerToken);