+++ CHANGELOG.md
@@ -9,10 +9,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 [unreleased changes details]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-5.0.6...HEAD
 
+## 5.0.14 - 2021-10-20
+
 ### Fixed
 
 - #2704 - Fixed issue with MCP report generation throwing an exception, and fixed some minor UI issues on AEM SDK (added BG color)
+- #2716 - Fixed issue with Shared Component Properties Bindings Values Provider facing lock contention
 - #2718 - Fixes CM Code Quality Pipeline failure caused by TestMarketoInterfaces and Jacoco instrumentation
+- #2713 - Marketo form/cloud config root missing
+- #2714 - Implemented shared and global component properties to work in experience fragments.
+- #2721 - Redirect Manager: Fix Broken UI in Cloud SDK
+- #2724 - Marketo proxy request support
 
 ## 5.0.12 - 2021-09-24